Add PlayerGameLog helper functions
Refactor PlayerGameLog mount method with helper functions.
